inheritance , extra result , cache path  and implementation of pcasting paper is added
