+++ codes/consumers.txt
@@ -19,7 +19,7 @@ c3:p3
 5
 5
 ----------
-c4:prod1
+c4:p3
 30
 200
 5
