+++ codes/faces.txt
@@ -40,6 +40,6 @@ c3:p3
 ----------
 f9
 3
-c4:prod1
+c4:p3
 cons2:prod2
 ---------
